+++ src/Command/Install/SetPermissionsCommand.php
@@ -45,6 +45,6 @@ class SetPermissionsCommand extends Command
      */
     public function execute(Arguments $args, ConsoleIo $io): void
     {
-        array_map(fn($path) => $this->folderChmod($io, $path), array_unique(Configure::read('WRITABLE_DIRS')));
+        array_map(fn(string $path): bool => $this->folderChmod($io, $path), array_unique(Configure::readOrFail('WRITABLE_DIRS')));
     }
 }