+++ src/TestSuite/TestCase.php
@@ -17,7 +17,6 @@ namespace MeTools\TestSuite;
 
 use Cake\Core\Configure;
 use Cake\ORM\Table;
-use Cake\ORM\TableRegistry;
 use Cake\TestSuite\TestCase as CakeTestCase;
 use Throwable;
 use Tools\Exceptionist;
@@ -83,8 +82,7 @@ abstract class TestCase extends CakeTestCase
      * Asserts a sql query string ends not with `$suffix`
      * @param string $suffix Suffix
      * @param string $sql Sql query string
-     * @param string $message The failure message that will be appended to the
-     *  generated message
+     * @param string $message The failure message that will be appended to the generated message
      * @return void
      * @since 2.20.7
      */
@@ -97,8 +95,7 @@ abstract class TestCase extends CakeTestCase
      * Asserts a sql query string ends with `$suffix`
      * @param string $suffix Suffix
      * @param string $sql Sql query string
-     * @param string $message The failure message that will be appended to the
-     *  generated message
+     * @param string $message The failure message that will be appended to the generated message
      * @return void
      * @since 2.20.7
      */
@@ -134,17 +131,13 @@ abstract class TestCase extends CakeTestCase
      * Get a table instance from the registry
      * @param string $alias The alias name you want to get
      * @param array $options The options you want to build the table with
-     * @return \Cake\ORM\Table|null
+     * @return \Cake\ORM\Table
      * @since 2.18.11
      */
-    protected function getTable(string $alias, array $options = []): ?Table
+    protected function getTable(string $alias, array $options = []): Table
     {
-        if ($alias === 'App' || (isset($options['className']) && !class_exists($options['className']))) {
-            return null;
-        }
+        $this->getTableLocator()->clear();
 
-        TableRegistry::getTableLocator()->clear();
-
-        return TableRegistry::getTableLocator()->get($alias, $options);
+        return $this->getTableLocator()->get($alias, $options);
     }
 }