+++ tests/TestCase/Command/Install/CreatePluginsLinksCommandTest.php
@@ -33,18 +33,19 @@ class CreatePluginsLinksCommandTest extends TestCase
      */
     public function testExecute(): void
     {
-        $clear = fn() => array_map(fn(string $path): bool => file_exists($path) && Filesystem::instance()->rmdirRecursive($path), [WWW_ROOT . 'me_tools', WWW_ROOT . 'test_plugin']);
+        $Filesystem = new Filesystem();
+        $clear = fn() => array_map(fn(string $path): bool => file_exists($path) && $Filesystem->rmdirRecursive($path), [WWW_ROOT . 'me_tools', WWW_ROOT . 'test_plugin']);
 
         $this->loadPlugins(['TestPlugin' => []]);
 
         $clear();
         $this->exec('me_tools.create_plugins_links -v');
-        $this->assertExitWithSuccess();
+        $this->assertExitSuccess();
         $this->assertOutputContains('Skipping plugin `Assets`. It does not have webroot folder');
         $this->assertOutputContains('For plugin: MeTools');
-        $this->assertOutputContains('Link `' . Filesystem::instance()->rtr(WWW_ROOT . 'me_tools') . '` has been created');
+        $this->assertOutputContains('Link `' . $Filesystem->rtr(WWW_ROOT . 'me_tools') . '` has been created');
         $this->assertOutputContains('For plugin: TestPlugin');
-        $this->assertOutputContains('Link `' . Filesystem::instance()->rtr(WWW_ROOT . 'test_plugin') . '` has been created');
+        $this->assertOutputContains('Link `' . $Filesystem->rtr(WWW_ROOT . 'test_plugin') . '` has been created');
         $this->assertErrorEmpty();
         $this->assertFileExists(WWW_ROOT . 'me_tools');
         $this->assertFileExists(WWW_ROOT . 'test_plugin');