+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 2.x branch
 ## 2.22 branch
+### 2.22.2
+* `assertExitWithError()` and `assertExitWithSuccess()` methods provided by `ConsoleIntegrationTestTrait` are now
+  deprecated. Use instead `assertExitError()` and `assertExitSuccess()`;
+* `MockTrait::getMockForComponent()` is deprecated. Use instead `createPartialMock()`;
+* `MockTrait::getMockForController()` is deprecated. Create instead a new instance of `Controller`;
+* many, small code improvements.
+
 ### 2.22.1
 * `Html::youtube()` and `BBCode::youtube()` methods can handle the starting second (`start` get parameter);
 * `BBCode::youtube()` method removes unnecessary `<p>` tags around the code;